requirements.txt
requests
beautifulsoup4
pandas
lxml
mysql-connector-python
jupyter
ipykernel
pymongo